Change return types to 'object' to allow for NoOp

+++ Interpreter.Tests/src/TestInterpreter.cs
@@ -12,30 +12,30 @@ namespace InterpreterPractice.Tests {
         [TestCase("3*2", ExpectedResult=6)]
         [TestCase("4/2", ExpectedResult=2)]
         public int Interpret_Calculation(string text) {
-            return new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
+            return (int)new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
         }
 
         [TestCase("1 + 2 * 3", ExpectedResult=7)]
         [TestCase("1 / 2 + 3", ExpectedResult=3)]
         [TestCase("4 * 3 / 2", ExpectedResult=6)]
         public int Parse_RespectsArithmeticPrecedence(string text) {
-            return new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
+            return (int)new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
         }
 
         [TestCase("(1 + 1) * 3", ExpectedResult=6)]
         [TestCase("2 + (1 + 3) * 3", ExpectedResult=14)]
         public int Parse_RespectParentheses(string text) {
-            return new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
+            return (int)new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
         }
 
         [TestCase("3 * (2 + (5 - 3))", ExpectedResult=12)]
         public int Parse_HandleNestedParentheses(string text) {
-            return new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
+            return (int)new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
         }
 
         [TestCase("5---2", ExpectedResult=3)]
         public int Parse_HandleUnaryOperator(string text) {
-            return new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
+            return (int)new Interpreter(new Parser(new Lexer(text).Next())).Interpret();
         }
     }
 }
